requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
twilio==6.59.1
urllib3==1.26.5
waitress==2.0.0
Werkzeug==2.0.1
WTForms==2.3.3
